+++ clients/spot/src/rest-api/types/klines-item.ts
@@ -19,9 +19,13 @@
  * Do not edit the class manually.
  */
 
+// May contain unused imports in some cases
+// @ts-ignore
+import type { KlinesItemInner } from './klines-item-inner';
+
 /**
  *
  * @export
  * @interface KlinesItem
  */
-export interface KlinesItem extends Array<string> {}
+export interface KlinesItem extends Array<KlinesItemInner> {}