+++ clients/spot/src/websocket-api/types/ui-klines-item.ts
@@ -19,9 +19,13 @@
  * Do not edit the class manually.
  */
 
+// May contain unused imports in some cases
+// @ts-ignore
+import type { KlinesItemInner } from './klines-item-inner';
+
 /**
  *
  * @export
  * @interface UiKlinesItem
  */
-export interface UiKlinesItem extends Array<string> {}
+export interface UiKlinesItem extends Array<KlinesItemInner> {}